requirements.txt
alembic==1.13.2
asgi-lifespan==2.*
asyncpg==0.29.0
aiosqlite==0.20.0
redis[hiredis]==4.6.0
python-barcode==0.15.1
bcrypt==4.2.0
fastapi[standard]==0.112.0
fastapi-mail==1.4.1
fastapi-cache2[redis]==0.2.2
pydantic==2.8.2
PyJWT==2.9.0
python-dotenv==1.0.0
PyYAML==6.0.1
sqlalchemy[asyncio]==2.0.31
starlette-admin==0.14.1
uvicorn==0.30.5
uvloop==0.19.0
httpx==0.27.0
pytest==8.3.3
pytest-asyncio==0.24.0